% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ameInfo.R
\docType{class}
\name{ameInfo-class}
\alias{ameInfo-class}
\alias{ameInfo}
\title{Class of ameInfo}
\description{
Class of ameInfo
}
\section{Slots}{

\describe{
\item{\code{final_score}}{The best score of the model}

\item{\code{converged}}{Logical. If TRUE, then the model was stopped before the maximum iteration was reached.}

\item{\code{num_iters}}{The number of interations run}

\item{\code{training_info}}{A data.frame containing training info over time.}
}}

