% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame-fit.R
\name{amem}
\alias{amem}
\title{Function for fitting adaptable mixed effects models}
\usage{
amem(formula, data, val_data = NULL, method = "xgboost",
  fixed_options = list(), min_iter = 10, max_iter = 20,
  stopping_criteria = 1e-05, keep_training_data = TRUE, chatty = TRUE,
  na_action = na.pass)
}
\arguments{
\item{formula}{As with \code{lme4::lmer}, a two-sided linear formula object describing both the fixed-effects and random-effects part of the model, with the response on the left of a ~ operator and the terms, separated by + operators, on the right. Random-effects terms are distinguished by vertical bars (|) separating expressions for design matrices from grouping factors.}

\item{data}{a data.frame containing the variables named in \code{formula}}

\item{val_data}{a data.frame, identical to data; If provided, early_stopping uses validation set.}

\item{method}{The method to use for fixed-effects -- currently not implemented.}

\item{fixed_options}{List of params to apply to fixed calculation.   See \code{xgb.train}}

\item{min_iter}{The minimum number of iterations for the optimization procedure (Default = 10)}

\item{max_iter}{The maximum number of iterations for the optimization procedure (Default = 20).}

\item{stopping_criteria}{Early stopping criteria, dependent on outcome. (Default = 1e-5)}

\item{keep_training_data}{Logical. If TRUE, training data is stored in the model object.}

\item{chatty}{Whether to be verbose or not (Default: TRUE)}

\item{na_action}{What to do with NAs (Default = \code{na.pass})}
}
\value{
an ameMod object
}
\description{
Function for fitting adaptable mixed effects models
}
\details{
Note that although the ame-fit funtion accepts formulas with complex random-slope effects, this functionality is experimental, and the slope effect of a given feature may not accurately represent the fixed-effect of that same feature. In addition, not every complex method of defining random effects has been fully tested.\code{max_iter} is used to define the value of \code{xgboost::nrounds} - if 'nrounds' is provided in the \code{fixed_options}, it will override the value of max_iter.
}
