% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame-tune.R
\name{amem.tune}
\alias{amem.tune}
\title{Bayesian optimization of amem model}
\usage{
amem.tune(formula, data, nfolds, val_data = NULL, tuning_params = NULL,
  seed = as.integer(Sys.time()))
}
\arguments{
\item{formula}{As with \code{lme4::lmer}, a two-sided linear formula object describing both the fixed-effects and random-effects part of the model, with the response on the left of a ~ operator and the terms, separated by + operators, on the right. Random-effects terms are distinguished by vertical bars (|) separating expressions for design matrices from grouping factors.}

\item{data}{a data.frame containing the variables named in \code{formula}}

\item{nfolds}{Number of folds for cross-validation (Required)}

\item{val_data}{a data.frame, identical to \code{data}.  Optional, and separate from tuning. If provided, nfolds is ignored.}

\item{tuning_params}{A list containing the parameters to tune, and the bounds of the search space for xgboost (see \code{?xgb.train} for descriptions. Currently accepts:
\itemize{
   \item \code{nrounds} Default c(20, 200)
   \item \code{max_depth} Default = c(2L, 6L)
   \item \code{min_child_weight} Default = c(1L, 10L)
   \item \code{eta} Default = c(.1, .5)
   \item \code{gamma} Default = 1 (not tuned).
   \item \code{colsample_bytree} Default = 1 (not tuned)
   \item \code{subsample} Default = 1 (not tuned)
}}

\item{seed}{Used for generating cross-validation folds (Default uses system time)}
}
\description{
Bayesian optimization of amem model
}
