% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ameMod.R
\docType{class}
\name{ameMod-class}
\alias{ameMod-class}
\alias{.ameMod}
\title{An S4 class of type ameMod}
\description{
An S4 class of type ameMod
}
\section{Slots}{

\describe{
\item{\code{call}}{The 'call' used to generate the model}

\item{\code{model_info}}{Information about the model training process}

\item{\code{fixef_model}}{The model representing the fixed effects}

\item{\code{ranef_model}}{An lme4 class object}

\item{\code{fixed_terms}}{A character vector of the column names used as fixed effects}

\item{\code{random_terms}}{A character vector of random terms structure in form 'Z | B'}

\item{\code{resid}}{Residuals of the model}

\item{\code{train_data}}{Data used to train the model.}

\item{\code{predictions}}{List containing training and validation set predictions}
}}

