% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{cb.amem}
\alias{cb.amem}
\title{Custom Callback for XGBOOST}
\usage{
cb.amem()
}
\description{
Custom Callback for XGBOOST
}
\details{
Handles the random-effects portion within xgboost iterations, returning both
The xgboost fixed model, as well as the
}
