% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{cb.early.stop.amem}
\alias{cb.early.stop.amem}
\title{Early Stopping with required minimum}
\usage{
cb.early.stop.amem(stopping_criteria, min_iter, maximize = FALSE,
  metric_name = NULL, verbose = FALSE)
}
\arguments{
\item{stopping_criteria}{The minimum change value for stopping}

\item{min_iter}{Minimum number of iterations}

\item{maximize}{Whether the function is attempting to maximize objective function (default is FALSE)}

\item{metric_name}{The name of the metric by which to evaluate model performance}

\item{verbose}{Logical. Whether to be chatty. Default FALSE.}
}
\description{
Modified version of xgboost::cb.early.stop() callback
Uses minimum iterations, and with nrounds used as 'maximum'
}
