% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_fixed_matrix}
\alias{create_fixed_matrix}
\title{Create fixed effect matrix}
\usage{
create_fixed_matrix(data, fixed_terms, outcome = NULL)
}
\arguments{
\item{data}{The set of data to use in creating the model.matrix and getting outcome info}

\item{fixed_terms}{Terms to use in fixed effect}

\item{outcome}{Column to use as outcome}
}
\value{
A list, FM, containing
\describe{
  \item{fixed_data}{the model matrix}
  \item{outcome}{a vector of outcome data}
  }
}
\description{
Create fixed effect matrix
}
