% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-internal-data.R
\name{create_internal_data}
\alias{create_internal_data}
\title{Create Internal Object}
\usage{
create_internal_data(dat, formula)
}
\arguments{
\item{dat}{The data object}

\item{formula}{The formula used to identify the fixed and random features}
}
\value{
Object of class xgb.DMatrix object, with additional attributes
\describe{
  \item{\code{rform}}{Formula for use in the random effect calculations}
  \item{\code{random_data}}{data.frame for use in the random-effect calculation}
  \item{\code{origLabel}}{The original outcome vector for the callback}
}
}
\description{
Create Internal Object
}
