% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GLL}
\alias{GLL}
\title{Generalized Log Likelihood function}
\usage{
GLL(epsil, b)
}
\arguments{
\item{epsil}{a vector of model residuals (Where epsilon = Y - Xb - Zu from trained models)}

\item{b}{The random-effect betas. Should be the same length as epsil}
}
\description{
Generalized Log Likelihood function
}
\details{
Note: Because \code{AMEM} uses \code{lme4::lmer} to generate random-effects betas with a 0-valued fixed effect formula (i.e., \code{outcome ~ 0 + (1 | random_effect)}), running \code{predict} on the lmerMod object provides the betas.
}
