% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-ameMod.R
\name{predict.ameMod}
\alias{predict.ameMod}
\title{Predict function for ameMod class}
\usage{
predict.ameMod(object, newdata, type_fixed, type_random, allow_new_levels,
  separate_preds, na_action, ...)
}
\arguments{
\item{object}{an ameMod object}

\item{newdata}{New data. If not provided, predictions will be generated from training data, if available}

\item{type_fixed}{Type of prediction for fixed effects ("raw" or "prob")}

\item{type_random}{Type of prediction for random effects ("response" or "link")}

\item{allow_new_levels}{Logical if new levels (or NA values) in newdata are allowed. If FALSE (default), such new values in newdata will trigger an error; if TRUE, then the prediction will use the unconditional (population-level) values for data with previously unobserved levels (or NAs).}

\item{separate_preds}{Logical. If TRUE, a data.frame containing separate fixed/random effects columns is produced. If FALSE, only the final combined prediction (as a vector) is returned}

\item{na_action}{function determining what should be done with missing values for fixed effects in newdata. The default is to predict NA: see na.pass}

\item{...}{Other parameters to pass to \code{caret::predict}}
}
\description{
Predict function for ameMod class
}
