% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{radon}
\alias{radon}
\title{Radon data}
\format{An object of class \code{data.frame} with 919 rows and 30 columns.}
\usage{
radon
}
\description{
A dataset containing 919 radon measurements with additional data
}
\keyword{datasets}
