% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ameInfo.R
\name{update_ameInfo_data}
\alias{update_ameInfo_data}
\title{Method to update ameInfo class with new data}
\usage{
update_ameInfo_data(object, GLL, TrainingMAE, TrainingR, HoldoutMAE, HoldoutR,
  Time_Sec)
}
\arguments{
\item{object}{an ameInfo object}

\item{GLL}{current Generalized Log Likelihood}

\item{TrainingMAE}{Current training MAE}

\item{TrainingR}{Correlation coefficient (R) of the training data predictions}

\item{HoldoutMAE}{Current holdout MAE (if applicable)}

\item{HoldoutR}{Correlation coefficient (R) of validation data}

\item{Time_Sec}{Time of current Iteration}
}
\description{
Method to update ameInfo class with new data
}
