% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ameMod.R
\name{varImp,ameMod-method}
\alias{varImp,ameMod-method}
\title{Variable importance for fixed effects measures}
\usage{
\S4method{varImp}{ameMod}(object)
}
\arguments{
\item{object}{amem model object}
}
\description{
Variable importance for fixed effects measures
}
